fix issue #1)  / 1.1 / [11:02pm]
https://leetcode.com/problems/average-salary-excluding-the-minimum-and-maximum-salary/
Have only done 5 easy algos or so. I believe this is new algo; if not new to me. 

Issue #1 is all of the issues pertaining to average-salary algorithm.
There will be multiple issues that could range quite variably. 

Using a:
int[ 1.1, 1.2, 1.3] Float system to organize and make easy reference of.
emboldened comments in the issue while signal which [ 1.1     ||     1.7 ] the commit would be referencing.